removed writes to cout for exrheader_fuzzer
